Report errors instead of tripping assertions (mypyc/mypyc#514)
Add an error reporting system and report errors instead of crashing
on unimplemented or unsupported features.

Also adds checks for some things we would silently miscompile before.

Error reporting is non-blocking. If we hit an error while compiling an
expression, we create a dummy temporary of the right type so code that
uses it can still be compiled. (Of course, the only reason to keep
compiling it is to see if it produces more errors.)