New semantic analyzer: Support multiple passes, typing imports (#6243)
Currently additional passes are only performed to resolve
"from import" statements and assignments. This allows
resolving indirect (exported) imports and NewType
definitions within import cycles.

Add a fixed maximum number of iterations that should always be
enough. If we hit the maximum, it's very likely that the remaining
unbound definitions are impossible to bind, or that there is a
definition cycle.

This also includes some fixes to support importing `typing`
in tests and adds support for `True` and `False` (and a few
other small changes).