New semantic analyzer: Fix redefinitions (#6257)
When we can't resolve a definition due to some missing dependencies,
we now don't record any additional definitions of the same name. This
is important since the first definition should always win. This applies
to things like assignments, function definitions, imports and class
definitions (anything that defines a name in the current scope).

Example of where things would go wrong:

```
from m import C  # Assume C can't be resolved yet

class C:  # Could be exposed, even though it's not the first def
    pass
```

Star imports are a bit special, since they could define arbitrary names
if the target module is incomplete. We use a special name `*` to
mark this. It disallows the definition of all names.

We don't track the scopes of definitions for simplicity. If definition
of `x` goes wrong, we also refuse to define it in any nested scope.
This may result in some extra deferrals, but otherwise shouldn't be
a problem.

This also refactors symbol table management. The key thing here is
that we want symbol table additions to go through a single function
so that we can skip them if they are defining something that was
already defined, but the previous definition couldn't be resolved. 

Some places still haven't been updated, but these are related to
features that aren't supported yet by the new analyzer. Also removed
some code related function redefinition, but again this isn't supported
yet and will have to be fixed anyway.