Categorize some stubgen tests as slow tests (#6340)
This allows pytest to parallelize tests more effectively and
seems to make the test suite faster by ~5 seconds or so on my
Linux desktop.

These can use typeshed so they are not fast tests.

Some of the StubgenPython tests are fast but others aren't. I'm
classifying them as slow since it looks like this is generally
better for scheduling, but it might be even better to split the
suite into fast and slow parts.