New semantic analyzer: conditional function definitions (#6429)
To semantically analyze conditional function definitions (and function
redefinitions) we add redefinitions to a symbol table with a mangled
name of form `name-redefinition[N]`. This is necessary since
functions are only analyzed if they can be accessed through a symbol
table.

I checked that this fixes most relevant test cases in
`check-functions.test`. Using a decorated function in a redefinition
is still not working perfectly. That is also partially implemented in
the old semantic analyzer. Conditionally defining an overloaded
function is also not supported. 

Fixes #6318.