Fix missing builtin aliases on Python 2 (and other versions) (#6441)
Some builtin aliases (like `typing.ChainMap`) are not available on some Python versions, or in some fixtures. However, if while they were first looked up while target namespace was incomplete, we put a `PlaceholderNode` in `typing`. This PR either doesn't put or removes this placeholder node if we didn't find the target after the target namespace is complete.

I didn't find a way to test this (since we don't have incremental `pythoneval` tests), but I double-checked this fixes the crash with actual Python 2 stubs.