Sort mypyc_targets in setup.py (#6569)
The mypy_mypyc-wheels Travis build is currently broken on linux
because of a mypyc bug that is dependent on the order of files passed
to mypyc. I am fixing the bug
(https://github.com/mypyc/mypyc/pull/539), but this was complicated by
files coming out of the file system APIs in an unspecified order.

Sort the targets passed to mypyc so that the mypyc compilation process
is deterministic.