Bump mypy pin (mypyc/mypyc#610)
This commit bumps the pinned version of mypy, mainly to verify the recent
changes to the semantics of the `Instance.final_value` have not broken
mypyc in any way. (It turned out mypyc was actually using the
`Var.final_value` field, which has the same name by coincidence.)

This commit also makes a few fixes to the tests to incoporate some of
the recent changes made in our error messages during the pycon sprints.