Make the runtime tests be invoked via an extension module (mypyc/mypyc#615)
Currently the runtime tests are built as a standalone binary linked against
libpython by using `python-config` to produce cflags and ldflags. This turns
out to be shockingly fragile and error-prone, since python-config location
and behavior doesn't really seem to be standardized. For example,
`python3-config`, which we default to using, doesn't exist in virtualenvs,
though a `python-config` does. So when run from a virtualenv, we actually
run the runtime tests against the system default Python 3, not the virtualenv's
Python. But we can't just switch to using `python-config` because that will
often point to Python 2.
Worse than all that, for reasons I have yet to fully understand I was never able
to get the python-config for dev versions on Travis to work. Using the python3-config
in the root python 3.8-dev install directory failed because it didn't include the library
path (??) and using the python-config in the virtualenv failed because it didn't include
`-lpython` (???) (or maybe vice versa).

Meanwhile building C extension modules inside of virtualenvs works pretty reliably,
so just switch to driving the runtime tests from a C extension module instead of from
a standalone binary. This means that the RT tests will run against the correct python
version on developer machines and we can turn on 3.8-dev tests in Travis.