Allow class-based TypedDicts in all python versions (#6971)
Entries can't be *declared* in versions that don't support type
annotations, of course, but class-based TypedDict still serves a
valuable purpose in merging TypedDicts together.

This error *only* triggers on empty TypedDicts, since any non-empty
ones would have failed earlier with a syntax error. So there is no
reason to prohibit it because we can handle that case totally fine.