New semantic analyzer: track line numbers of placeholders (#6983)
Line numbers are used sometimes to decide when a name should be looked
up from an outer scope, so they should be set for placeholders.

This actually breaks some forward references to assignment-based named
tuples. The reason is that we don't set `becomes_typeinfo=True` for
the placeholders. I'm going to open a separate issue about this, as
it's an existing issue and this change only exposes it.

Fixes #6949.