New analyzer: Fix Any kinds for special forms (#7148)
Fixes https://github.com/python/mypy/issues/7118

Previously left hand sides of special forms were inferred as `TypeOfAny.from_error`. Note that _total_ number of expressions in one of the tests is now smaller because of the subtle logic in visiting definitions of special forms, but I don't think it is import to keep it one-to-one with old analyser.