Add beginnings of error code support (#7267)
This PR adds a foundation for error codes and implements a few error
codes. It also adds support for `# type: ignore[code1, ...]` which 
ignores only specific error codes on a line.

Only a few errors include interesting error codes at this point. I'll add 
support for more error codes in additional PRs. Most errors will implicitly 
fall back to a `misc` error code.

Error codes are only shown if `--show-error-codes` is used.

The error codes look like this in mypy output:

```
t.py:3: error: "str" has no attribute "trim"  [attr-defined]
```

Error codes are intended to be short but human-readable. The name
of an error code refers to the check that produces this error. In the above
example we generate a "no attribute" error when we check whether an 
attribute is defined.

Work towards #7239.