[mypyc] Support separate compilation (#7636)
This adds support for separate compilation to mypyc.

A separate argument is added to mypycify. If True, mypyc places every
module its own separate shared library. Otherwise it can take a list of groups
of source files that should be placed in a shared library together.

The shared libraries communication with each other using the C API's intended
mechanism for communicating between C extension modules: Capsules.
Each library creates a table of pointers to all of its exported data and stores a pointer
to it in a capsule stored as a module attribute. When a library is loaded, it
loads the linking table capsules from all its dependencies and copies them
into a local copy of the table. (To eliminate the need for a pointer indirection when
accessing it.)

This adds a test mode that will run all multi-module run tests in separate compilation mode.
I also manually tested mypy itself compiled in separate compilation mode.

This supports a limited form of incremental compilation already: only modules
that changed or had a header they depend on change will be recompiled by the
C compiler. The entire project still needs to go through the entire mypy/mypyc
front-end and middle-end, however. I expect to have a PR that adds support for
hooking into incremental mode next week.

This is progress on mypyc/mypyc#682.