[mypyc] Allow dots in group names (#7941)
Single-module groups will be given a dotted group name equal to the
module name by default.

A dot in a group name will put the group C extension at that
path. Generated C could will also be placed in a directory
corresponding to the full dotted name (still under the build
directory).

The direct motivation here is that bazel wants artifacts to be under
the appropriate directory, but it is cleaner in general I think.