Refine parent type when narrowing "lookup" expressions (#7917)
This diff adds support for the following pattern:

```python
from typing import Enum, List
from enum import Enum

class Key(Enum):
    A = 1
    B = 2

class Foo:
    key: Literal[Key.A]
    blah: List[int]

class Bar:
    key: Literal[Key.B]
    something: List[str]

x: Union[Foo, Bar]
if x.key is Key.A:
    reveal_type(x)  # Revealed type is 'Foo'
else:
    reveal_type(x)  # Revealed type is 'Bar'
```

In short, when we do `x.key is Key.A`, we "propagate" the information
we discovered about `x.key` up to refine the type of `x`.

We perform this propagation only when `x` is a Union and only when we
are doing member or index lookups into instances, typeddicts,
namedtuples, and tuples. For indexing operations, we have one additional
limitation: we *must* use a literal expression in order for narrowing
to work at all. Using Literal types or Final instances won't work;
See https://github.com/python/mypy/issues/7905 for more details.

To put it another way, this adds support for tagged unions, I guess.

This more or less resolves https://github.com/python/mypy/issues/7344.
We currently don't have support for narrowing based on string or int
literals, but that's a separate issue and should be resolved by
https://github.com/python/mypy/pull/7169 (which I resumed work
on earlier this week).