[suggest] Fix inference from collection literals (#7979)
This is another take at #7810 which was reverted because it caused
internal trouble.

Do this by ignoring constraints with Any when the Any comes from the
suggestion engine. The Any context causes assigned types to be Any
overaggressively.  Restricting it only to Anys from the suggestion
engine avoids the issues seen last time. Doing it at a fine-grained
level in constraints allows us to do better refining things like `List[Any]`.