Fix regression in container check logic (#8232)
This PR fixes the crash reported in https://github.com/python/mypy/issues/8230,
by replacing the 'pass' with the 'continue', as suggested.

However, it does *not* fix the underlying root cause -- I don't think
I actually understand the relevant pieces of code enough to feel
confident volunteering a fix. So, I settled for just fixing the
regression.

Basically, it seems this bug is due to how we try inferring the type
of the lambda in multiple passes to resolve the types. We pencil in an
ErasedType during the first pass -- and then subsequently crash when
attempting to type check the body during that pass. I'll leave more
details about this in the linked issue.