Fix an incremental crash bug caused by ad-hoc intersection (#8381)
Ad-hoc intersection was generating type names with dots in them, which
are interpreted as module separators. I fixed this by making the names
worse and just using the bare name of the class.

There are other options that could be pursued:
* Replacing "." with some other character and then either switching it
  back when displayed or just printing it. One option that I was very
  tempted by was to use U+2024 ONE DOT LEADER.
* Compute the intersection names on the fly when formatting.

I decided to do the most obvious and the least tricky thing because
this isn't important.