Make IRBuilder not inherit from LowLevelIRBuilder. (#8426)
Instead we make IRBuilder do all of its building through a
LowLevelIRBuilder. A new LowLevelIRBuilder is created for each
function that is compiled (which lets LowLevelIRBuilder lose its stack
of block lists and environments and only has one of each).

A collection of the most commonly used methods in LowLevelIRBuilder
are given passthrough methods in IRBuilder for convenience.

Work on mypyc/mypyc#714.